third_party/WebKit/LayoutTests/http/tests/htmlimports/redirect-cross-origin-cross-same-expected.txt

CONSOLE ERROR: Imported resource from origin 'http://127.0.0.1:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'null' is therefore not allowed access.
This is a testharness.js-based test.
PASS Accessing to a same origin import including CORS domain redirect 
Harness: the test ran to completion.
